requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.2
nltk==3.8.1
numpy>=1.21.0
google-cloud-aiplatform>=1.36.0
google-generativeai>=0.3.0
plotly>=5.13.0
scikit-learn>=1.0.2
networkx>=2.8.4
rich>=12.0.0 